Integrals.jl

Questo modulo implementa un metodo di integrazione finita di polinomi.

Interfacce principali

II(P::LAR, alpha::Int, beta::Int, gamma::Int, signedInt=false)::Float64

III(P::LAR, alpha::Int, beta::Int, gamma::Int, signedInt::Bool=false)::Float64

surface(P::LAR, signedInt::Bool=false)::Float64

volume(P::LAR, signedInt::Bool=false)::Float64

firstMoment(P::LAR)::Array{Float64,1}

secondMoment(P::Lar.LAR)::Array{Float64,1}

inertiaProduct(P::LAR)::Array{Float64,1}

centroid(P::LAR)::Array{Float64,1}

inertiaMoment(P::LAR)::Array{Float64,1}

Documentazione funzioni

Integrals.M — Function

M(alpha::Int, beta::Int)::Float64

Module for integration of polynomials over 3D volumes and surfaces

Integrals.TT — Function

TT(tau::Array{Float64,2}, alpha::Int, beta::Int, gamma::Int, signedInt::Bool=false)

The main integration routine

Integrals.II — Function

II(P::Lar.LAR, alpha::Int, beta::Int, gamma::Int, signedInt=false)

Basic integration function on 2D plane.

Example unit 3D triangle

julia> V = [0.0 1.0 0.0; 0.0 0.0 1.0; 0.0 0.0 0.0]
3×3 Array{Float64,2}:
 0.0  1.0  0.0
 0.0  0.0  1.0
 0.0  0.0  0.0
 
julia> FV = [[1,2,3]]
1-element Array{Array{Int64,1},1}:
 [1, 2, 3]
 
julia> P = V,FV
([0.0 1.0 0.0; 0.0 0.0 1.0; 0.0 0.0 0.0], Array{Int64,1}[[1, 2, 3]])

julia> Integrals.II(P, 0,0,0)
0.5

Integrals.III — Function

III(P::LAR, alpha::Int, beta::Int, gamma::Int, signedInt::Bool=false)::Float64

Basic integration function on 3D space.

Example # unit 3D tetrahedron

julia> V = [0.0 1.0 0.0 0.0; 0.0 0.0 1.0 0.0; 0.0 0.0 0.0 1.0]
3×4 Array{Float64,2}:
 0.0  1.0  0.0  0.0
 0.0  0.0  1.0  0.0
 0.0  0.0  0.0  1.0

julia> FV = [[1, 2, 4], [1, 3, 2], [4, 3, 1], [2, 3, 4]]
4-element Array{Array{Int64,1},1}:
 [1, 2, 4]
 [1, 3, 2]
 [4, 3, 1]
 [2, 3, 4]

julia> P = V,FV
([0.0 1.0 0.0 0.0; 0.0 0.0 1.0 0.0; 0.0 0.0 0.0 1.0], 
Array{Int64,1}[[1, 2, 4], [1, 3, 2], [4, 3, 1], [2, 3, 4]])

julia> Integrals.III(P, 0,0,0)
0.16666666666666674

Integrals.surface — Function

surface(P::Lar.LAR, signedInt::Bool=false)::Float64

surface integral on polyhedron P.

Example unit 3D triangle

julia> V = [0.0 1.0 0.0; 0.0 0.0 1.0; 0.0 0.0 0.0]
3×3 Array{Float64,2}:
 0.0  1.0  0.0
 0.0  0.0  1.0
 0.0  0.0  0.0
 
julia> FV = [[1,2,3]]
1-element Array{Array{Int64,1},1}:
 [1, 2, 3]
 
julia> P = V,FV
([0.0 1.0 0.0; 0.0 0.0 1.0; 0.0 0.0 0.0], Array{Int64,1}[[1, 2, 3]])

julia> Integrals.surface(P)
0.5

Integrals.volume — Function

volume(P::LAR, signedInt::Bool=false)::Float64

volume integral on polyhedron P.

Example # unit 3D tetrahedron

julia> V = [0.0 1.0 0.0 0.0; 0.0 0.0 1.0 0.0; 0.0 0.0 0.0 1.0]
3×4 Array{Float64,2}:
 0.0  1.0  0.0  0.0
 0.0  0.0  1.0  0.0
 0.0  0.0  0.0  1.0

julia> FV = [[1, 2, 4], [1, 3, 2], [4, 3, 1], [2, 3, 4]]
4-element Array{Array{Int64,1},1}:
 [1, 2, 4]
 [1, 3, 2]
 [4, 3, 1]
 [2, 3, 4]

julia> P = V,FV
([0.0 1.0 0.0 0.0; 0.0 0.0 1.0 0.0; 0.0 0.0 0.0 1.0], 
Array{Int64,1}[[1, 2, 4], [1, 3, 2], [4, 3, 1], [2, 3, 4]])

julia> Integrals.volume(P)
0.16666666666666674

Integrals.firstMoment — Function

firstMoment(P::Lar.LAR)::Array{Float64,1}

First moments as terms of the Euler tensor. Remember that the integration algorithm is a boundary integration. Hence the model must be a boundary model. In this case, a 2-complex of triangles.

Example # unit 3D tetrahedron

julia> V = [0.0 1.0 0.0 0.0; 0.0 0.0 1.0 0.0; 0.0 0.0 0.0 1.0];

julia> FV = [[1, 2, 4], [1, 3, 2], [4, 3, 1], [2, 3, 4]];

julia> P = V,FV;

julia> Integrals.firstMoment(P)
3-element Array{Float64,1}:
 0.0416667
 0.0416667
 0.0416667

Integrals.secondMoment — Function

secondMoment(P::Lar.LAR)::Array{Float64,1}

Second moments as terms of the Euler tensor.

Example # unit 3D tetrahedron

julia> V = [0.0 1.0 0.0 0.0; 0.0 0.0 1.0 0.0; 0.0 0.0 0.0 1.0];

julia> FV = [[1, 2, 4], [1, 3, 2], [4, 3, 1], [2, 3, 4]];

julia> P = V,FV;

julia> Integrals.secondMoment(P)
3-element Array{Float64,1}:
 0.0166667
 0.0166667
 0.0166667

Integrals.inertiaProduct — Function

inertiaProduct(P::Lar.LAR)::Array{Float64,1}

Inertia products as terms of the Euler tensor.

Example # unit 3D tetrahedron

julia> V = [0.0 1.0 0.0 0.0; 0.0 0.0 1.0 0.0; 0.0 0.0 0.0 1.0];

julia> FV = [[1, 2, 4], [1, 3, 2], [4, 3, 1], [2, 3, 4]];

julia> P = V,FV;

julia> Integrals.inertiaProduct(P)
3-element Array{Float64,1}:
 0.00833333
 0.00833333
 0.00833333

Integrals.centroid — Function

centroid(P::Lar.LAR)::Array{Float64,1}

Barycenter or centroid of polyhedron P.

Example # unit 3D tetrahedron

julia> V = [0.0 1.0 0.0 0.0; 0.0 0.0 1.0 0.0; 0.0 0.0 0.0 1.0];

julia> FV = [[1, 2, 4], [1, 3, 2], [4, 3, 1], [2, 3, 4]];

julia> P = V,FV;

julia> Integrals.centroid(P)
3-element Array{Float64,1}:
 0.25
 0.25
 0.25

Integrals.inertiaMoment — Function

inertiaMoment(P::Lar.LAR)::Array{Float64,1}

Inertia moments of polyhedron P.

Example # unit 3D tetrahedron

julia> V = [0.0 1.0 0.0 0.0; 0.0 0.0 1.0 0.0; 0.0 0.0 0.0 1.0];

julia> FV = [[1, 2, 4], [1, 3, 2], [4, 3, 1], [2, 3, 4]];

julia> P = V,FV;

julia> Integrals.inertiaMoment(P)
3-element Array{Float64,1}:
 0.0333333
 0.0333333
 0.0333333