Integral Solver Algorithms

The following algorithms are available:

QuadGKJL: Uses QuadGK.jl. Requires nout=1 and batch=0, in-place is not allowed.

HCubatureJL: Uses HCubature.jl. Requires batch=0.

VEGAS: Uses MonteCarloIntegration.jl. Requires nout=1. Works only for >1-dimensional integrations.

VEGASMC: Uses MCIntegration.jl. Requires using MCIntegration. Doesn't support batching.

CubatureJLh: h-Cubature from Cubature.jl. Requires using Cubature.

CubatureJLp: p-Cubature from Cubature.jl. Requires using Cubature.

CubaVegas: Vegas from Cuba.jl. Requires using Cuba, nout=1.

CubaSUAVE: SUAVE from Cuba.jl. Requires using Cuba.

CubaDivonne: Divonne from Cuba.jl. Requires using Cuba. Works only for >1-dimensional integrations.

CubaCuhre: Cuhre from Cuba.jl. Requires using Cuba. Works only for >1-dimensional integrations.

GaussLegendre: Uses Gauss-Legendre quadrature with nodes and weights from FastGaussQuadrature.jl.

QuadratureRule: Accepts a user-defined function that returns nodes and weights.

ArblibJL: real- and complex-valued univariate integration of holomorphic and meromorphic functions from Arblib.jl. Requires using Arblib.

Integrals.QuadGKJL — Type

QuadGKJL(; order = 7, norm=norm)

One-dimensional Gauss-Kronrod integration from QuadGK.jl. This method also takes the optional arguments order and norm. Which are the order of the integration rule and the norm for calculating the error, respectively

References

@article{laurie1997calculation, title={Calculation of Gauss-Kronrod quadrature rules}, author={Laurie, Dirk}, journal={Mathematics of Computation}, volume={66}, number={219}, pages={1133–1145}, year={1997} }

Integrals.HCubatureJL — Type

HCubatureJL(; norm=norm, initdiv=1)

Multidimensional "h-adaptive" integration from HCubature.jl. This method also takes the optional arguments initdiv and norm. Which are the initial number of segments each dimension of the integration domain is divided into, and the norm for calculating the error, respectively.

References

@article{genz1980remarks, title={Remarks on algorithm 006: An adaptive algorithm for numerical integration over an N-dimensional rectangular region}, author={Genz, Alan C and Malik, Aftab Ahmad}, journal={Journal of Computational and Applied mathematics}, volume={6}, number={4}, pages={295–302}, year={1980}, publisher={Elsevier} }

Integrals.CubatureJLp — Type

CubatureJLp()

Multidimensional p-adaptive integration from Cubature.jl. This method is based on repeatedly doubling the degree of the cubature rules, until convergence is achieved. The used cubature rule is a tensor product of Clenshaw–Curtis quadrature rules. error_norm specifies the convergence criterion for vector valued integrands. Defaults to Cubature.INDIVIDUAL, other options are Cubature.PAIRED, Cubature.L1, Cubature.L2, or Cubature.LINF.

Integrals.CubatureJLh — Type

CubatureJLh()

Multidimensional h-adaptive integration from Cubature.jl. error_norm specifies the convergence criterion for vector valued integrands. Defaults to Cubature.INDIVIDUAL, other options are Cubature.PAIRED, Cubature.L1, Cubature.L2, or Cubature.LINF.

References

@article{genz1980remarks, title={Remarks on algorithm 006: An adaptive algorithm for numerical integration over an N-dimensional rectangular region}, author={Genz, Alan C and Malik, Aftab Ahmad}, journal={Journal of Computational and Applied mathematics}, volume={6}, number={4}, pages={295–302}, year={1980}, publisher={Elsevier} }

Integrals.VEGAS — Type

VEGAS(; nbins = 100, ncalls = 1000, debug=false)

Multidimensional adaptive Monte Carlo integration from MonteCarloIntegration.jl. Importance sampling is used to reduce variance. This method also takes three optional arguments nbins, ncalls and debug which are the initial number of bins each dimension of the integration domain is divided into, the number of function calls per iteration of the algorithm, and whether debug info should be printed, respectively.

Limitations

This algorithm can only integrate Float64-valued functions

References

@article{lepage1978new, title={A new algorithm for adaptive multidimensional integration}, author={Lepage, G Peter}, journal={Journal of Computational Physics}, volume={27}, number={2}, pages={192–203}, year={1978}, publisher={Elsevier} }

Integrals.VEGASMC — Type

VEGASMC(; kws...)

Markov-chain based Vegas algorithm from MCIntegration.jl

Refer to MCIntegration.integrate for documentation on the keywords, which are passed directly to the solver with a set of defaults that works for conforming integrands.

Integrals.CubaVegas — Type

CubaVegas()

Multidimensional adaptive Monte Carlo integration from Cuba.jl. Importance sampling is used to reduce variance.

References

@article{lepage1978new, title={A new algorithm for adaptive multidimensional integration}, author={Lepage, G Peter}, journal={Journal of Computational Physics}, volume={27}, number={2}, pages={192–203}, year={1978}, publisher={Elsevier} }

Integrals.CubaSUAVE — Type

CubaSUAVE()

Multidimensional adaptive Monte Carlo integration from Cuba.jl. Suave stands for subregion-adaptive VEGAS. Importance sampling and subdivision are thus used to reduce variance.

References

@article{hahn2005cuba, title={Cuba—a library for multidimensional numerical integration}, author={Hahn, Thomas}, journal={Computer Physics Communications}, volume={168}, number={2}, pages={78–95}, year={2005}, publisher={Elsevier} }

Integrals.CubaDivonne — Type

CubaDivonne()

Multidimensional adaptive Monte Carlo integration from Cuba.jl. Stratified sampling is used to reduce variance.

References

@article{friedman1981nested, title={A nested partitioning procedure for numerical multiple integration}, author={Friedman, Jerome H and Wright, Margaret H}, journal={ACM Transactions on Mathematical Software (TOMS)}, volume={7}, number={1}, pages={76–92}, year={1981}, publisher={ACM New York, NY, USA} }

Integrals.CubaCuhre — Type

CubaCuhre()

Multidimensional h-adaptive integration from Cuba.jl.

References

@article{berntsen1991adaptive, title={An adaptive algorithm for the approximate calculation of multiple integrals}, author={Berntsen, Jarle and Espelid, Terje O and Genz, Alan}, journal={ACM Transactions on Mathematical Software (TOMS)}, volume={17}, number={4}, pages={437–451}, year={1991}, publisher={ACM New York, NY, USA} }

Integrals.GaussLegendre — Type

GaussLegendre{C, N, W}

Struct for evaluating an integral via (composite) Gauss-Legendre quadrature. The field C will be true if subintervals > 1, and false otherwise.

The fields nodes::N and weights::W are defined by nodes, weights = gausslegendre(n) for a given number of nodes n.

The field subintervals::Int64 = 1 (with default value 1) defines the number of intervals to partition the original interval of integration [a, b] into, splitting it into [xⱼ, xⱼ₊₁] for j = 1,…,subintervals, where xⱼ = a + (j-1)h and h = (b-a)/subintervals. Gauss-Legendre quadrature is then applied on each subinterval. For example, if [a, b] = [-1, 1] and subintervals = 2, then Gauss-Legendre quadrature will be applied separately on [-1, 0] and [0, 1], summing the two results.

Integrals.QuadratureRule — Type

QuadratureRule(q; n=250)

Algorithm to construct and evaluate a quadrature rule q of n points computed from the inputs as x, w = q(n). It assumes the nodes and weights are for the standard interval [-1, 1]^d in d dimensions, and rescales the nodes to the specific hypercube being solved. The nodes x may be scalars in 1d or vectors in arbitrary dimensions, and the weights w must be scalar. The algorithm computes the quadrature rule sum(w .* f.(x)) and the caller must check that the result is converged with respect to n.

Integrals.ArblibJL — Type

ArblibJL(; check_analytic=false, take_prec=false, warn_on_no_convergence=false, opts=C_NULL)

One-dimensional adaptive Gauss-Legendre integration using rigorous error bounds and precision ball arithmetic. Generally this assumes the integrand is holomorphic or meromorphic, which is the user's responsibility to verify. The result of the integral is not guaranteed to satisfy the requested tolerances, however the result is guaranteed to be within the error estimate.

Arblib.jl only supports integration of univariate real- and complex-valued functions with both inplace and out-of-place forms. See their documentation for additional details the algorithm arguments and on implementing high-precision integrands. Additionally, the error estimate is included in the return value of the integral, representing a ball.